fix: floorf → floor in HSVtoRGB to prevent sector-boundary hue shifts
floorf takes a float argument; passing a double h/60 value
implicitly narrows to 23-bit mantissa before flooring. Near HSV
sector boundaries (multiples of 60°), the narrowing can round
a value like 2.9999997 up to 3.0f, flipping to the wrong sector
and producing a completely different output hue. This manifested
as spectrum mode showing blue instead of green on Windows where
the float rounding differed from Linux/macOS.

floor() keeps the full 52-bit double precision through the
operation, giving consistent sector assignment across platforms.

+++ xyscope-shared.h
@@ -199,7 +199,7 @@ static inline void HSVtoRGB(double *r, double *g, double *b,
         h -= 360.0;
 
     h /= 60;
-    i = (int)floorf(h);
+    i = (int)floor(h);
     f = h - i;
     p = v * (1 - s);
     q = v * (1 - s * f);
